backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
minio==7.2.0
cryptography==41.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
aiofiles==23.2.1
PyPDF2==3.0.1
pillow==10.1.0
pydicom==2.4.3
pandas==2.1.3
numpy==1.26.2
python-magic==0.4.27
email-validator==2.1.0
httpx==0.25.2
prometheus-client==0.19.0
celery==5.3.4
flower==2.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3